fix(product): clean up auth and mobility review issues
